feat: type check struct definitions.
This commit implements type checking on struct definitions.

To accomplish this, a dependency graph is built for locally defined structs and
cycles are detected and a diagnostic is emitted for recursively defined
structs.

Otherwise, the type is calculated for the struct, which is done in topological
order to resolve type name references.

This includes a refactor to move code related to a V1 AST out of `scope.rs` and
into `scope/v1.rs`.

Also fixed a bug in the signature of `write_json` in the standard library: it
should return `File` and not `Union`.
