+++ wdl-analysis/tests/analysis/recursive-structs/foo.wdl
@@ -1,0 +1,5 @@
+version 1.1
+
+struct Foo {
+    Foo f
+}
