+++ wdl-analysis/tests/analysis/unknown-type/foo.wdl
@@ -1,0 +1,5 @@
+version 1.2
+
+struct X {
+    Int x
+}
